Commit initial doubly linked list and tests
Functions to insert, remove and count doubly linked-list elements and a
test bench to test the implementation. The functions are built to take
lists and elements and return status so that NULLs and invariant
violations can be caught, e.g. the prev and next pointer of a
non-inserted element should point to that element.

In addition, list elements only contain next and prev members. The user
is meant to use the CONTAINER_OF macro to get a pointer to the outer
struct that would contain a list element. CONTAINER_OF is implemented in
test_list.c, its usage is demonstrated and it has its own unit test.

ARRAY_SIZE is also implemented, demonstrated and unit tested.

Finally, Both CONTAINER_OF and ARRAY_SIZE should be in separate files,
but in the interest of file minimization, I've put them into
test_list.c. I note this because I wrote this to demonstrate good-ish
implementation practices. If something could be better, please let me
know.  For instance, I've used Doxygen to document the functions as
opposed to http://www.sphinx-doc.org/en/stable/domains.html. What are
your thoughts?